Remove rsleevi@chromium.org from third_party/modp_b64/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I65823f9ef263faefe36473ec29d2348308409ed1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3502091
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977219}
NOKEYCHECK=True
GitOrigin-RevId: 2fa997dafa32f59ba5bcab4ceb34fcadbf32356e